#354 moved all http client related code to utils
