+++ agent/discovery/discovery.go
@@ -90,7 +90,7 @@ func client() error {
 }
 
 func fingerprint() string {
-	client := config.GetClient(config.Management.Allowinsecure, 5)
+	client := utils.GetClient(config.Management.Allowinsecure, 5)
 	resp, err := client.Get("https://10.10.10.1:8443/rest/v1/security/keyman/getpublickeyfingerprint")
 	if err == nil {
 		defer utils.Close(resp)
@@ -127,7 +127,7 @@ func save(ip string) {
 }
 
 func getKey() []byte {
-	client := config.GetClient(config.Management.Allowinsecure, 5)
+	client := utils.GetClient(config.Management.Allowinsecure, 5)
 	resp, err := client.Get("https://" + config.Management.Host + ":" + config.Management.Port + config.Management.RestPublicKey)
 
 	if err == nil {
