+++ cli/import.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"net/http"
 	"os"
 	"strings"
 	"time"
@@ -56,7 +55,7 @@ type metainfo struct {
 }
 
 // getTemplateInfo retrieves the id of a template on global repository, id of the latest template version will be returned
-func getTemplateInfo(t *templ, kurjun *http.Client, token string) {
+func getTemplateInfo(t *templ, token string) {
 
 	//get template info by name from Kurjun
 	url := config.CDN.Kurjun + "/template/info?name=" + t.name + "&token=" + token
@@ -64,6 +63,7 @@ func getTemplateInfo(t *templ, kurjun *http.Client, token string) {
 		url += "&owner=" + t.owner[0]
 	}
 
+	kurjun := utils.GetClient(config.CDN.Allowinsecure, 15)
 	response, err := kurjun.Get(url)
 	log.Check(log.ErrorLevel, "Retrieving template info, get: "+url, err)
 	defer utils.Close(response)
@@ -93,11 +93,12 @@ func getTemplateInfo(t *templ, kurjun *http.Client, token string) {
 }
 
 // idToName retrieves template name from global repository by passed id string
-func idToName(t *templ, id string, kurjun *http.Client, token string) {
+func idToName(t *templ, id string, token string) {
 	//Since only kurjun knows template's ID, we cannot define if we have template already installed in system by ID as we do it by name, so unreachable kurjun in this case is a deadend for us
 	//To omit this issue we should add ID into template config and use this ID as a "primary key" to any request
 	url := config.CDN.Kurjun + "/template/info?id=" + id + "&token=" + token
 
+	kurjun := utils.GetClient(config.CDN.Allowinsecure, 15)
 	response, err := kurjun.Get(url)
 	log.Check(log.ErrorLevel, "Retrieving template info, get: "+url, err)
 	defer utils.Close(response)
@@ -165,7 +166,7 @@ func download(t templ, token string) (bool, error) {
 	}
 	defer out.Close()
 
-	client := config.GetClientForUploadDownload()
+	client := utils.GetClientForUploadDownload()
 
 	url := config.CDN.Kurjun + "/template/download?id=" + t.id + "&token=" + token
 
@@ -438,20 +439,18 @@ func LxcImport(name, token string, local bool, auxDepList ...string) {
 }
 
 func fetchTemplateMetadata(t *templ, token string) {
-	kurjun, err := config.CheckKurjun()
-
-	log.Check(log.ErrorLevel, "Connecting to Kurjun", err)
+	config.CheckCDN()
 
 	//obtain name and owner from Kurjun
 	if id := strings.Split(t.name, "id:"); len(id) > 1 {
-		idToName(t, id[1], kurjun, token)
+		idToName(t, id[1], token)
 	} else if line := strings.Split(t.name, "/"); len(line) > 1 {
 		t.name = line[1]
 		t.owner = append(t.owner, line[0])
 	}
 
 	//obtain full template info from Kurjun by name & owner to get latest version
-	getTemplateInfo(t, kurjun, token)
+	getTemplateInfo(t, token)
 
 	log.Info("Version: " + t.version)
